Update price to decimal

+++ src/MarketAssistant/Program.cs
@@ -80,7 +80,7 @@ static void ProductsMenu(Dictionary<string, (DateOnly DateOfExpiry, int Availabl
         Console.ReadLine();
     }
 
-    static void DeleteProductAction(Dictionary<string, (DateOnly DateOfExpiry, int AvailableAmount, int SoldAmount, int Price)>? products)
+    static void DeleteProductAction(Dictionary<string, (DateOnly DateOfExpiry, int AvailableAmount, int SoldAmount, decimal Price)>? products)
     {
         var deleteProductMenuItems = new List<(int Id, string Name)>
     {
